requirements.txt
python-dotenv
feedparser
python-dateutil
requests 
requests_oauthlib